+++ HISTORY.rst
@@ -6,7 +6,7 @@ History
 0.X.X (XXXX-XX-XX)
 ------------------
 * Update buildweights service to add support for regridding to domain file. Not backwards compatible. (PR#67, @dgergel)
-
+* Add downscaling service. Currently support BCSD spatial disaggregation as implemented in scikit-downscale. (PR#65, @dgergel)
 
 0.2.0 (2021-04-23)
 ------------------